opencti-worker/src/requirements.txt
git+https://github.com/OpenCTI-Platform/client-python@master
opentelemetry-api==1.20.0
opentelemetry-sdk==1.20.0
opentelemetry-exporter-prometheus==1.12.0rc1
requests~=2.31.0
google-auth==2.23.3